classifier/requirements.txt
numpy
onnxruntime
onnxruntime_gpu
pylsl
scipy
websocket_client